Init
Init push unstable already know some bugs. Just pushing this as a back up right now. Goal will be simple. Show stocks in relative realtime in a way that looks good, is OS independent, relatively efficient, and as a whole looks pleasing to the eye.

+ basic implementation

- header doesn't show yet

Notes
Currently have the scaling to work regardless of terminal size and IMPORTANTLY updates every print to the terminal size.